Squashed initial commit of the modular approach
