+++ Assets/Editor/UI/AssetHandlers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 908843777c8a7e746b54776368e5ecc3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
