+++ Assets/Editor/UI/AssetHandlers/DefaultAssetHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 532907903ca992649b7dc4a99796a07e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
