+++ Assets/Editor/UI/AssetHandlers/ImageSignAssetHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cb4d83a572bca3248ae7ac98e399d252
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
