+++ Assets/Editor/UI/AssetHandlers/SignAssetHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 72b14558e32f2e94eaa17a75eaba8eed
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
