+++ Assets/Editor/UI/AssetHandlers/TrainAssetHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 32b3ee7025f4c0249a6eca5a41df1295
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
